HDPI-5181: Replace counter-claim placeholder with counterclaim documents upload page
