--- src/main/assets/locales/cy/respondToClaim/doYouWantToUploadCounterclaimFiles.json
@@ -1,14 +1,0 @@
-{
-  "pageTitle": "cy Do you want to upload files?",
-  "caption": "cy Make a counterclaim",
-  "question": "cy Do you want to upload files to support your counterclaim?",
-  "options": {
-    "yes": "cy Yes",
-    "no": "cy No"
-  },
-  "errors": {
-    "wantsToUploadCounterClaimDocs": {
-      "required": "cy Select if you want to upload files to support your counterclaim"
-    }
-  }
-}
