+++ src/main/assets/locales/en/respondToClaim/counterClaim.json
@@ -1,1 +1,13 @@
-{}
+{
+  "pageTitle": "Upload counterclaim files",
+  "caption": "Make a counterclaim",
+  "heading": "Upload documents (optional)",
+  "guidanceText": "Upload any documents that you think are relevant to your counterclaim.",
+  "beforeUploadHeading": "Before you upload your documents",
+  "beforeUploadText": "Give your document a name that explains what it is. For example, 'Receipt of payment from Jane Smith'",
+  "uploadLabel": "Upload a file",
+  "filesAddedHeading": "Uploaded files",
+  "uploadButton": "Upload file",
+  "deleteButton": "Remove",
+  "noFilesUploaded": "No files uploaded"
+}
