--- src/main/assets/locales/en/respondToClaim/doYouWantToUploadCounterclaimFiles.json
@@ -1,14 +1,0 @@
-{
-  "pageTitle": "Do you want to upload files?",
-  "caption": "Make a counterclaim",
-  "question": "Do you want to upload files to support your counterclaim?",
-  "options": {
-    "yes": "Yes",
-    "no": "No"
-  },
-  "errors": {
-    "wantsToUploadCounterClaimDocs": {
-      "required": "Select if you want to upload files to support your counterclaim"
-    }
-  }
-}
