+++ src/main/steps/respond-to-claim/counter-claim/counterClaim.njk
@@ -1,16 +1,40 @@
 {% extends "stepsTemplate.njk" %}
-{% from "macros/stepButtons.njk" import stepButtons %}
 {% from "govuk/components/error-summary/macro.njk" import govukErrorSummary %}
+{% from "macros/stepButtons.njk" import stepButtons %}
 {% from "macros/csrf.njk" import csrfProtection %}
-{% block pageTitle %}Counterclaim &ndash; HM Courts &amp; Tribunals Service &ndash; GOV.UK{% endblock %}
+{% from "macros/fileUploadMoj.njk" import fileUploadMoj %}
+
+{% block pageTitle %}
+  {{ pageTitle }} &ndash; HM Courts &amp; Tribunals Service &ndash; GOV.UK
+{% endblock %}
 
 {% block mainContent %}
   {% if errorSummary %}
     {{ govukErrorSummary(errorSummary) }}
   {% endif %}
-  <h1 class="govuk-heading-l">Counterclaim (placeholder)</h1>
-  <p class="govuk-body">This is a placeholder for the counterclaim step.</p>
-  <form method="post" action="{{ url }}">
+
+  <span class="govuk-caption-xl">{{ caption }}</span>
+  {% if heading %}
+    <h1 class="govuk-heading-l">{{ heading }}</h1>
+  {% endif %}
+
+  <p class="govuk-body">{{ guidanceText }}</p>
+
+  <h2 class="govuk-heading-s">{{ beforeUploadHeading }}</h2>
+  <p class="govuk-body">{{ beforeUploadText }}</p>
+
+  <form method="post" action="{{ url }}" novalidate>
+    {% for field in fields %}
+      {% if field.componentType == 'fileUpload' %}
+        {{ fileUploadMoj(
+          field.component,
+          field.component.uploadButtonText,
+          field.component.filesAddedHeading,
+          field.component.value
+        ) }}
+      {% endif %}
+    {% endfor %}
+
     {{ stepButtons(saveAndContinue, saveForLater) }}
     {{ csrfProtection(csrfToken) }}
   </form>
