--- src/main/steps/respond-to-claim/do-you-want-to-upload-counterclaim-files/doYouWantToUploadCounterclaimFiles.njk
@@ -1,28 +1,0 @@
-{% extends "stepsTemplate.njk" %}
-{% from "govuk/components/error-summary/macro.njk" import govukErrorSummary %}
-{% from "govuk/components/radios/macro.njk" import govukRadios %}
-{% from "macros/stepButtons.njk" import stepButtons %}
-{% from "macros/csrf.njk" import csrfProtection %}
-
-{% block pageTitle %}
-  {{ pageTitle }} &ndash; HM Courts &amp; Tribunals Service &ndash; GOV.UK
-{% endblock %}
-
-{% block mainContent %}
-  {% if errorSummary %}
-    {{ govukErrorSummary(errorSummary) }}
-  {% endif %}
-
-  <span class="govuk-caption-xl">{{ caption }}</span>
-
-  <form method="post" action="{{ url }}" novalidate>
-    {% for field in fields %}
-      {% if field.componentType == 'radios' %}
-        {{ govukRadios(field.component) }}
-      {% endif %}
-    {% endfor %}
-
-    {{ stepButtons(saveAndContinue, saveForLater) }}
-    {{ csrfProtection(csrfToken) }}
-  </form>
-{% endblock %}
